Move Pulumi storage account and key to context
