+++ data_safe_haven/config/config.py
@@ -3,10 +3,9 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Any, ClassVar
+from typing import Any
 
 import yaml
-from azure.keyvault.keys import KeyVaultKey
 from pydantic import (
     BaseModel,
     Field,
@@ -59,8 +58,6 @@ class ConfigSectionAzure(BaseModel, validate_assignment=True):
 
 
 class ConfigSectionPulumi(BaseModel, validate_assignment=True):
-    storage_container_name: ClassVar[str] = "pulumi"
-    encryption_key_name: ClassVar[str] = "pulumi-encryption-key"
     stacks: dict[str, str] = Field(..., default_factory=dict[str, str])
 
 
@@ -239,8 +236,6 @@ class Config(BaseModel, validate_assignment=True):
     )
     tags: ConfigSectionTags = Field(..., exclude=True)
 
-    _pulumi_encryption_key = None
-
     def __init__(self, context: Context, **kwargs: dict[Any, Any]):
         tags = ConfigSectionTags(context)
         super().__init__(context=context, tags=tags, **kwargs)
@@ -257,22 +252,6 @@ class Config(BaseModel, validate_assignment=True):
     @property
     def work_directory(self) -> Path:
         return self.context.work_directory
-
-    @property
-    def pulumi_encryption_key(self) -> KeyVaultKey:
-        if not self._pulumi_encryption_key:
-            azure_api = AzureApi(subscription_name=self.context.subscription_name)
-            self._pulumi_encryption_key = azure_api.get_keyvault_key(
-                key_name=self.pulumi.encryption_key_name,
-                key_vault_name=self.context.key_vault_name,
-            )
-        return self._pulumi_encryption_key
-
-    @property
-    def pulumi_encryption_key_version(self) -> str:
-        """ID for the Pulumi encryption key"""
-        key_id: str = self.pulumi_encryption_key.id
-        return key_id.split("/")[-1]
 
     @property
     def sre_names(self) -> list[str]:
