+++ data_safe_haven/config/context_settings.py
@@ -9,6 +9,7 @@ from pathlib import Path
 from typing import ClassVar
 
 import yaml
+from azure.keyvault.keys import KeyVaultKey
 from pydantic import BaseModel, Field, ValidationError, model_validator
 from yaml import YAMLError
 
@@ -16,6 +17,7 @@ from data_safe_haven.exceptions import (
     DataSafeHavenConfigError,
     DataSafeHavenParameterError,
 )
+from data_safe_haven.external import AzureApi
 from data_safe_haven.functions import alphanumeric
 from data_safe_haven.utility import LoggingSingleton, config_dir
 from data_safe_haven.utility.annotated_types import (
@@ -35,6 +37,10 @@ class Context(BaseModel, validate_assignment=True):
     name: str
     subscription_name: AzureLongName
     storage_container_name: ClassVar[str] = "config"
+    pulumi_storage_container_name: ClassVar[str] = "pulumi"
+    pulumi_encryption_key_name: ClassVar[str] = "pulumi-encryption-key"
+
+    _pulumi_encryption_key = None
 
     @property
     def shm_name(self) -> str:
@@ -67,6 +73,30 @@ class Context(BaseModel, validate_assignment=True):
 
     def to_yaml(self) -> str:
         return yaml.dump(self.model_dump(), indent=2)
+
+    @property
+    def pulumi_backend_url(self) -> str:
+        return f"azblob://{self.pulumi_storage_container_name}"
+
+    @property
+    def pulumi_encryption_key(self) -> KeyVaultKey:
+        if not self._pulumi_encryption_key:
+            azure_api = AzureApi(subscription_name=self.subscription_name)
+            self._pulumi_encryption_key = azure_api.get_keyvault_key(
+                key_name=self.pulumi_encryption_key_name,
+                key_vault_name=self.key_vault_name,
+            )
+        return self._pulumi_encryption_key
+
+    @property
+    def pulumi_encryption_key_version(self) -> str:
+        """ID for the Pulumi encryption key"""
+        key_id: str = self.pulumi_encryption_key.id
+        return key_id.split("/")[-1]
+
+    @property
+    def pulumi_secrets_provider_url(self) -> str:
+        return f"azurekeyvault://{self.key_vault_name}.vault.azure.net/keys/{self.pulumi_encryption_key_name}/{self.pulumi_encryption_key_version}"
 
 
 class ContextSettings(BaseModel, validate_assignment=True):
