+++ data_safe_haven/infrastructure/stack_manager.py
@@ -47,7 +47,7 @@ class PulumiAccount:
                 "AZURE_STORAGE_ACCOUNT": self.cfg.context.storage_account_name,
                 "AZURE_STORAGE_KEY": str(backend_storage_account_keys[0].value),
                 "AZURE_KEYVAULT_AUTH_VIA_CLI": "true",
-                "PULUMI_BACKEND_URL": f"azblob://{self.cfg.pulumi.storage_container_name}",
+                "PULUMI_BACKEND_URL": self.cfg.context.pulumi_backend_url,
             }
         return self.env_
 
@@ -101,7 +101,7 @@ class StackManager:
                     stack_name=self.stack_name,
                     program=self.program.run,
                     opts=automation.LocalWorkspaceOptions(
-                        secrets_provider=f"azurekeyvault://{self.cfg.context.key_vault_name}.vault.azure.net/keys/{self.cfg.pulumi.encryption_key_name}/{self.cfg.pulumi_encryption_key_version}",
+                        secrets_provider=self.cfg.context.pulumi_secrets_provider_url,
                         work_dir=str(self.work_dir),
                         env_vars=self.account.env,
                     ),
@@ -216,7 +216,7 @@ class StackManager:
                     blob_name=f".pulumi/stacks/{self.project_name}/{stack_backup_name}",
                     resource_group_name=self.cfg.context.resource_group_name,
                     storage_account_name=self.cfg.context.storage_account_name,
-                    storage_container_name=self.cfg.pulumi.storage_container_name,
+                    storage_container_name=self.cfg.context.pulumi_storage_container_name,
                 )
             except DataSafeHavenAzureError as exc:
                 if "blob does not exist" in str(exc):
